+++ frontend/src/lib/auth/activateUser.ts
