+++ frontend/src/lib/auth/registerUser.ts
