+++ frontend/src/lib/auth/verifyToken.ts
