+++ frontend/src/lib/chat/getChatFirstMsg.ts
